requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.1
Deprecated==1.2.13
Django==4.0.3
django-ckeditor==6.3.2
django-js-asset==2.0.0
django-oauth-toolkit==2.0.0
djangorestframework==3.13.1
idna==3.3
jwcrypto==1.2
mysqlclient==2.1.0
oauthlib==3.2.0
Pillow==9.1.0
pycparser==2.21
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
wrapt==1.14.1
